Fazendo o parse do JSON em asynctask
